The files for my Capstone
I would first open my Databricks file as that has an explanation of my project in it but these are all of the files for my Final Project for DS3002! It also includes screenshots from the platforms I used, Azure, MongoDB, and Azure Data Studio. I did not include a screenshot of the Databricks file but that can be opened using Databricks.